Move repository files
