+++ apps/web/src/app/api/article/new/route.ts
@@ -6,7 +6,7 @@ import { prisma, Prisma } from "@labatory/db";
 import { getServerSession } from "next-auth";
 import { authOptions } from "@/lib/auth";
 
-import { CreateArticleCore } from "@/repository/db/article";
+import { CreateArticleCore } from "@/repository/db/article/article";
 import { buildCreateArticleCtx } from "@/app/api/_util/createArticleCtx";
 
 type Body = {
