+++ apps/web/src/app/univ/actions.ts
@@ -1,6 +1,6 @@
 "use server";
 
-import { CreateTag, UpdateTag } from "@/repository/db/tag";
+import { CreateTag, UpdateTag } from "@/repository/db/article/tag";
 import { prisma } from "@labatory/db";
 import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
